+++ SrcBxxh/l0commsgccl/commsgccl.h
@@ -160,19 +160,19 @@ typedef struct com_sensor_status //
 //MSG_ID_COM_INIT
 typedef struct msg_struct_com_init //
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_com_init_t;
 
 //MSG_ID_COM_INIT_FB
 typedef struct msg_struct_com_init_fb //
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_com_init_fb_t;
 
 //MSG_ID_COM_RESTART
 typedef struct  msg_struct_com_restart //
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_com_restart_t;
 
 //MSG_ID_COM_TIME_OUT
@@ -180,31 +180,31 @@ typedef struct  msg_struct_com_time_out //
 {
 	UINT32 timeId;
 	UINT8  timeRes;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_com_time_out_t;
 
 //MSG_ID_COM_STOP
 typedef struct  msg_struct_com_stop //
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_com_stop_t;
 
 //MSG_ID_COM_HEART_BEAT
 typedef struct  msg_struct_com_heart_beat //
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_com_heart_beat_t;
 
 //MSG_ID_COM_HEART_BEAT_FB
 typedef struct  msg_struct_com_heart_beat_fb //
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_com_heart_beat_fb_t;
 
 //VMFO消息定义
 typedef struct msg_struct_vmfo_1s_period_timtout
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_vmfo_1s_period_timtout_t;
 
 //DIDO
@@ -213,7 +213,7 @@ typedef struct msg_struct_dido_ccl_sensor_status_rep
 {
 	UINT8 cmdid;
 	com_sensor_status_t sensor;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_dido_ccl_sensor_status_rep_t;
 
 //MSG_ID_DIDO_CCL_EVENT_LOCK_TRIGGER
@@ -221,7 +221,7 @@ typedef struct msg_struct_dido_ccl_event_lock_trigger
 {
 	UINT8 cmdid;
 	UINT8 lockid;  //指示是哪一个锁触发的
-	UINT8 length;
+	UINT16 length;
 }msg_struct_dido_ccl_event_lock_trigger_t;
 
 //MSG_ID_DIDO_CCL_EVENT_FAULT_TRIGGER
@@ -229,40 +229,40 @@ typedef struct msg_struct_dido_ccl_event_fault_trigger
 {
 	UINT8 cmdid;
 	com_sensor_status_t sensor;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_dido_ccl_event_fault_trigger_t;
 
 //MSG_ID_DIDO_CCL_EVENT_STATUS_UPDATE
 typedef struct msg_struct_dido_ccl_status_update
 {
 	UINT8 cmdid;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_dido_ccl_status_update_t;
 
 //MSG_ID_DIDO_CCL_DOOR_OPEN_EVENT
 typedef struct msg_struct_dido_door_open_event
 {
 	UINT8 cmdid;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_dido_ccl_door_open_event_t;
 
 //MSG_ID_DIDO_CCL_LOCK_C_DOOR_C_EVENT
 typedef struct msg_struct_dido_ccl_lock_c_door_c_event
 {
 	UINT8 cmdid;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_dido_ccl_lock_c_door_c_event_t;
 
 //SPS消息定义
 typedef struct msg_struct_spsvirgo_l2frame_send
 {
 	UINT8 data[MAX_IHU_MSG_BODY_LENGTH-1];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_spsvirgo_l2frame_send_t;
 typedef struct msg_struct_spsvirgo_l2frame_rcv
 {
 	UINT8 data[MAX_IHU_MSG_BODY_LENGTH-1];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_spsvirgo_l2frame_rcv_t;
 
 //MSG_ID_SPS_CCL_CLOUD_FB
@@ -270,7 +270,7 @@ typedef struct msg_struct_spsvirgo_ccl_cloud_fb
 {
 	UINT8 cmdid;
 	UINT8 authResult;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_spsvirgo_ccl_cloud_fb_t;
 
 //MSG_ID_SPS_CCL_SENSOR_STATUS_RESP
@@ -278,25 +278,25 @@ typedef struct msg_struct_sps_ccl_sensor_status_rep
 {
 	UINT8 cmdid;
 	com_sensor_status_t sensor;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_sps_ccl_sensor_status_rep_t;
 
 //MSG_ID_SPS_CCL_EVENT_REPORT_CFM
 typedef struct msg_struct_sps_ccl_event_report_cfm
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_sps_ccl_event_report_cfm_t;
 
 //MSG_ID_SPS_CCL_FAULT_REPORT_CFM
 typedef struct msg_struct_sps_ccl_fault_report_cfm
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_sps_ccl_fault_report_cfm_t;
 
 //MSG_ID_SPS_CCL_CLOSE_REPORT_CFM
 typedef struct msg_struct_sps_ccl_close_report_cfm
 {
-	UINT8 length;
+	UINT16 length;
 }msg_struct_sps_ccl_close_report_cfm_t;
 
 //SPI消息定义
@@ -304,26 +304,26 @@ typedef struct msg_struct_sps_ccl_close_report_cfm
 typedef struct msg_struct_spileo_l2frame_send
 {
 	UINT8 data[MAX_IHU_MSG_BODY_LENGTH-1];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_spileo_l2frame_send_t;
 
 //MSG_ID_SPI_L2FRAME_RCV
 typedef struct msg_struct_spileo_l2frame_rcv
 {
 	UINT8 data[MAX_IHU_MSG_BODY_LENGTH-1];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_spileo_l2frame_rcv_t;
 
 //I2C消息定义
 typedef struct msg_struct_i2caries_l2frame_send
 {
 	UINT8 data[MAX_IHU_MSG_BODY_LENGTH-1];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_i2caries_l2frame_send_t;
 typedef struct msg_struct_i2caries_l2frame_rcv
 {
 	UINT8 data[MAX_IHU_MSG_BODY_LENGTH-1];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_i2caries_l2frame_rcv_t;
 
 //MSG_ID_I2C_CCL_SENSOR_STATUS_RESP
@@ -331,19 +331,19 @@ typedef struct msg_struct_i2c_ccl_sensor_status_rep
 {
 	UINT8 cmdid;
 	com_sensor_status_t sensor;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_i2c_ccl_sensor_status_rep_t;
 
 //CAN消息定义
 typedef struct msg_struct_canvela_l2frame_send
 {
 	UINT8 data[MAX_IHU_MSG_BODY_LENGTH-1];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_canvela_l2frame_send_t;
 typedef struct msg_struct_canvela_l2frame_rcv
 {
 	UINT8 data[MAX_IHU_MSG_BODY_LENGTH-1];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_canvela_l2frame_rcv_t;
 
 //DCMI
@@ -352,7 +352,7 @@ typedef struct msg_struct_dcmi_ccl_sensor_status_rep
 {
 	UINT8 cmdid;
 	com_sensor_status_t sensor;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_dcmi_ccl_sensor_status_rep_t;
 
 //CCL
@@ -362,14 +362,14 @@ typedef struct msg_struct_ccl_sps_open_auth_inq
 {
 	UINT8 cmdid;
 	UINT8 dataBuf[IHU_CCL_BH_CTRL_CMD_BUF_LEN];
-	UINT8 length;
+	UINT16 length;
 }msg_struct_ccl_sps_open_auth_inq;
 
 //MSG_ID_CCL_COM_SENSOR_STATUS_REQ
 typedef struct msg_struct_ccl_com_sensor_status_req
 {
 	UINT8 cmdid;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_ccl_com_sensor_status_req_t;
 
 //MSG_ID_CCL_COM_CTRL_CMD
@@ -378,7 +378,7 @@ typedef struct msg_struct_ccl_com_ctrl_cmd
 	UINT8 workmode;
 	UINT8 cmdid;
 	UINT8 lockid;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_ccl_com_ctrl_cmd_t;
 
 //MSG_ID_CCL_SPS_EVENT_REPORT_SEND
@@ -386,7 +386,7 @@ typedef struct msg_struct_ccl_sps_event_report_send
 {
 	UINT8 cmdid;
 	com_sensor_status_t sensor;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_ccl_sps_event_report_send_t;
 
 //MSG_ID_CCL_SPS_FAULT_REPORT_SEND
@@ -394,7 +394,7 @@ typedef struct msg_struct_ccl_sps_fault_report_send
 {
 	UINT8 cause;
 	com_sensor_status_t sensor;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_ccl_sps_fault_report_send_t;
 #define IHU_CCL_FAULT_CAUSE_SENSOR_WARNING 1
 #define IHU_CCL_FAULT_CAUSE_CLOSE_DOOR_TIME_OUT 2
@@ -403,7 +403,7 @@ typedef struct msg_struct_ccl_sps_fault_report_send
 typedef struct msg_struct_ccl_sps_close_report_send
 {
 	UINT8 cause;
-	UINT8 length;
+	UINT16 length;
 }msg_struct_ccl_sps_close_report_send_t;
 #define IHU_CCL_CLOSE_DOOR_NORMAL 1
 #define IHU_CCL_CLOSE_DOOR_NOT_YET_OPEN 2
